multi: move client conn creation func to mailbox
In this commit, the MailboxRPCConnection function is renamed and moved
out of the core package and into the mailbox package.
